Modify migrations, add API for trucks and addresses

+++ migrations/1647609049-CreateTableAddresses.js
@@ -2,12 +2,11 @@ exports.up = async (sql) => {
   await sql`
 		CREATE TABLE addresses (
 			id integer PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
-			street_number integer (60) NOT NULL,
+			street_number integer NOT NULL,
 			street_name varchar (100) NOT NULL,
 			zipcode varchar (30) NOT NULL,
 			country varchar (100) NOT NULL,
 			company_name varchar (100) NOT NULL
-
 		);
 	`;
 };
